+++ linux/flutter/generated_plugins.cmake
@@ -6,6 +6,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   desktop_webview_auth
   platform_device_id_linux
   screen_retriever
+  sentry_flutter
   url_launcher_linux
   window_manager
 )